+++ SPARC_4550/SPARC_4550.X/Timers.h
@@ -8,10 +8,11 @@
 #ifndef TIMERS_H
 #define	TIMERS_H
 #include <xc.h>
+#include <stdint.h>
 void tmr0Init(void); //Se inicializan los registros para que el TMR0 sea contador de eventos
 void tmr1Init(void); //Se inicializan los registros para que el TMR1 sea contador de eventos
-void setNumPasosX(unsigned int numPasosX); //Funcion para poner a contar al TMR0 los pasos requeridos
-void setNumPasosY(unsigned int numPasosY); //Fucnion para poner a contar al TMR1los pasos requeridos
+void setNumPasosX(uint16_t numPasosX); //Funcion para poner a contar al TMR0 los pasos requeridos
+void setNumPasosY(uint16_t numPasosY); //Fucnion para poner a contar al TMR1los pasos requeridos
 
 #endif	/* TIMERS_H */
